Make LoadRepositoryForTest importable from varats
