+++ tests/utils/test_git_util.py
@@ -4,13 +4,13 @@ from pathlib import Path
 
 from benchbuild.utils.revision_ranges import RevisionRange
 
-from tests.test_utils import LoadRepositoryForTest
 from varats.project.project_util import (
     get_local_project_git,
     get_local_project_git_path,
     BinaryType,
 )
 from varats.projects.discover_projects import initialize_projects
+from varats.utils.git_commands import LoadRepository
 from varats.utils.git_util import (
     ChurnConfig,
     CommitRepoPair,
@@ -124,7 +124,7 @@ class TestGitInteractionHelpers(unittest.TestCase):
         """Check if correct submodule commit is retrieved."""
         repo_head = FullCommitHash("cb15dfa4b2d7fba0d50e87b49f979c7f996b8ebc")
 
-        with LoadRepositoryForTest("grep", repo_head.to_short_commit_hash()):
+        with LoadRepository("grep", repo_head.to_short_commit_hash()):
             submodule_head = get_submodule_head("grep", "gnulib", repo_head)
             self.assertEqual(
                 submodule_head,
@@ -135,7 +135,7 @@ class TestGitInteractionHelpers(unittest.TestCase):
         """Check if correct main repo commit is retrieved."""
         repo_head = FullCommitHash("cb15dfa4b2d7fba0d50e87b49f979c7f996b8ebc")
 
-        with LoadRepositoryForTest("grep", repo_head.to_short_commit_hash()):
+        with LoadRepository("grep", repo_head.to_short_commit_hash()):
             submodule_head = get_submodule_head("grep", "grep", repo_head)
             self.assertEqual(submodule_head, repo_head)
